Update of course Selection.
Added option to view courses not specific to one program(e.g. BCEE) or
that do not have a course code starting by the program name.
Also, now when viewing course details and going back, it will directly
load the course list of the program for which the course detail was
being viewed (instead of having to click on program all the time to view
the course list).